% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_selection.R
\name{before_after}
\alias{before_after}
\title{One direction date sub setter}
\usage{
before_after(direction, date)
}
\arguments{
\item{date}{}
}
\value{
Returns subset dates for function of "before_after"
}
\description{
One direction date sub setter
}
\examples{
before_after("leq","2021-08-30")
}
