% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_date_format.R, R/data.R
\docType{data}
\name{covid_data}
\alias{covid_data}
\title{Covid Date formatter}
\format{
An object of class \code{data.frame} with 499 rows and 22 columns.

A data frame with 499 obs. of 22 variables
}
\source{
Original data from <https://data.sandiegocounty.gov/dataset/COVID-19-Statistics-San-Diego-County/uvug-znjd/data>
}
\usage{
covid_data

covid_data
}
\arguments{
\item{data}{}
}
\value{
returns date column as date
}
\description{
An example for Group 8 R final.
This is a modified version of San Diego counties Covid dataset from 2000-2022
fixed for missing values and formatted dates
}
\examples{
converted_data$Date <- as.Date(converted_data$Date)
}
\keyword{datasets}
