% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_selection.R
\name{double_dates}
\alias{double_dates}
\title{Subset vector of dates between from one date to another}
\usage{
double_dates(earlier_date, later_date)
}
\arguments{
\item{later_date}{}
}
\value{
returns the subset of one date earlier all the way to another date
later
}
\description{
Subset vector of dates between from one date to another
}
\examples{
double_dates("2020-08-21","2021-09-21")
}
